portfolio-cms/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-filter==24.1
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.0
Pillow==10.1.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
